Corrections

+++ src/main.rs
@@ -26,7 +26,6 @@ use std::error::Error;
 use std::fs;
 use std::io;
 use std::path::Path;
-use std::process::{exit, ExitCode};
 use std::time::Duration;
 use zeroize::Zeroize;
 const CACHEFORMAT: &str = "%Y-%m-%d-%H-%M-%S";
@@ -353,7 +352,7 @@ fn rocket() -> _ {
     let mut key = fs::read_to_string(config.itkey).unwrap();
     let rsakey = getprivatekey(&key).unwrap();
     key.zeroize();
-    let port: u8 = u8::try_from(config.port).unwrap();
+    let port: u16 = u16::try_from(config.port).unwrap();
     let config = Config {
         issuer_hash,
         issuer_name_hash,
